+++ pypeit/edgetrace.py
@@ -387,6 +387,8 @@ class EdgeTraceSet(DataContainer):
             A floating-point array with the coefficients (offset, scale) of the
             x-correlation between RIGHT edges predicted by the optical model and the
             ones traced on the image.
+        maskfile (:obj:`str`):
+            Full path to the file used to extract slit-mask information
     """
 
     master_type = 'Edges'
@@ -398,7 +400,7 @@ class EdgeTraceSet(DataContainer):
     bitmask = EdgeTraceBitMask()
     """BitMask instance."""
 
-    version = '1.0.0'
+    version = '1.0.1'
     """DataContainer datamodel version."""
 
     output_float_dtype = np.float32
@@ -418,6 +420,9 @@ class EdgeTraceSet(DataContainer):
                                   descr='Sobel-filtered image used to detect edges'),
                  'traceid': dict(otype=np.ndarray, atype=(int, np.integer),
                                  descr='ID number for the edge traces.  Negative and positive '
+                                       'IDs are for, respectively, left and right edges.'),
+                 'maskdef_id': dict(otype=np.ndarray, atype=(int, np.integer),
+                                 descr='slitmask ID number for the edge traces. '
                                        'IDs are for, respectively, left and right edges.'),
                  'orderid': dict(otype=np.ndarray, atype=(int, np.integer),
                                  descr='For echelle spectrographs, this is the order ID number '
@@ -479,6 +484,8 @@ class EdgeTraceSet(DataContainer):
         self.dispname = spectrograph.dispname           # Spectrograph disperser
         self.par = par                                  # Parameters used for slit edge tracing
         self.qa_path = qa_path                          # Directory for QA plots
+        self.maskdef_id = None                          # Slit ID number from slit-mask design
+                                                        # matched to traced slits
 
         # NOTE: This means that, no matter what, every instance of
         # EdgeTraceSet should have a sobelsig attribute that is *not*
@@ -505,13 +512,13 @@ class EdgeTraceSet(DataContainer):
         self.log = None                 # Log of methods applied
         self.master_key = None          # Calibration key for master frame
         self.master_dir = None          # Directory for Master frames
-        self.maskdef_id = None          # Slit ID number from slit-mask design matched to traced slits
         self.omodel_bspat = None        # Left edges predicted by the optical model (before x-correlation)
         self.omodel_tspat = None        # Right edges predicted by the optical model (before x-correlation)
         self.coeff_b = None             # Coefficients of the x-correlation between LEFT edges predicted
                                         # by the optical model and traced on the image.
         self.coeff_t = None             # Coefficients of the x-correlation between LEFT edges predicted
                                         # by the optical model and traced on the image.
+        self.maskfile = None            # File used to slurp in slit-mask design
 
     def _reinit_trace_data(self):
         """
@@ -536,6 +543,7 @@ class EdgeTraceSet(DataContainer):
         self.omodel_tspat = None
         self.coeff_b = None
         self.coeff_t = None
+        self.maskfile = None
 
     @property
     def ntrace(self):
@@ -577,16 +585,18 @@ class EdgeTraceSet(DataContainer):
                                  description='Spatial pixel coordinate for left edge'),
                     table.Column(name='TRACERPIX', dtype=float, length=length,
                                  description='Spatial pixel coordinate for right edge'),
+                    table.Column(name='SPAT_ID', dtype=int, length=length,
+                                 description='ID Number assigned by the pypeline to each slit'),
                     table.Column(name='SLITID', dtype=int, length=length,
-                                 description='Slit ID Number'),
+                                 description='Slit ID Number from slit-mask design'),
                     # table.Column(name='SLITLFOC', dtype=float, length=length,
                     #              description='Left edge of the slit in mm at the focal plane'),
                     # table.Column(name='SLITRFOC', dtype=float, length=length,
                     #              description='Right edge of the slit in mm at the focal plane'),
                     table.Column(name='SLITLOPT', dtype=float, length=length,
-                                description='Left edge of the slit in pixel from optical model before x-correlation'),
+                                description='Left edge of the slit in pixel predicted by optical model'),
                     table.Column(name='SLITROPT', dtype=float, length=length,
-                                description='Right edge of the slit in pixel from optical model before x-correlation'),
+                                description='Right edge of the slit in pixel predicted optical model'),
                     table.Column(name='SLITRA', dtype=float, length=length,
                                  description='Right ascension of the slit center (deg)'),
                     table.Column(name='SLITDEC', dtype=float, length=length,
@@ -625,10 +635,16 @@ class EdgeTraceSet(DataContainer):
                                  description='Right ascension of the object (deg)'),
                     table.Column(name='OBJDEC', dtype=float, length=length,
                                  description='Declination of the object (deg)'),
+                    table.Column(name='OBJNAME', dtype='<U32', length=length,
+                                 description='Object name assigned by the observer'),
                     table.Column(name='SLITID', dtype=int, length=length,
                                  description='Slit ID Number'),
-                    table.Column(name='SLITINDX', dtype=int, length=length,
-                                 description='Row index of relevant slit in the design table')
+                    table.Column(name='OBJ_TOPDIST', dtype=float, length=length,
+                                 description='Projected position of the object w.r.t. the top of the slit (arcsec)'),
+                    table.Column(name='OBJ_BOTDIST', dtype=float, length=length,
+                                 description='Projected position of the object w.r.t. the bottom of the slit (arcsec)'),
+                    table.Column(name='TRACEID', dtype=int, length=length,
+                                 description='Row index that matches TRACEID in the design table')
                            ])
 
     def rectify(self, flux, bpm=None, extract_width=None, mask_threshold=0.5, side='left'):
@@ -1334,11 +1350,16 @@ class EdgeTraceSet(DataContainer):
             nslits = np.amax(traceid)   # Only used if synced is True
             synced = self.is_synced
             slit_ids = None
+            maskdef_ids = None
             if synced:
                 _trc = cen if fit is None else fit
                 half = _trc.shape[0] // 2
                 slit_ids = ((_trc[half, gpm & is_left]
                              + _trc[half, gpm & is_right]) / 2.).astype(int)
+                if self.maskdef_id is not None:
+                    maskdef_ids = self.maskdef_id[gpm & self.is_left]
+                    maskdef_ids[maskdef_ids == -99] = self.maskdef_id[gpm & self.is_right][maskdef_ids == -99]
+
         else:
             # Use the provided SlitTraceSet
             _include_error = False
@@ -1359,6 +1380,7 @@ class EdgeTraceSet(DataContainer):
             # known, but this functionality appears to be never used? I
             # think slits needs its own show method.
             slit_ids = slits.slitord_id
+            maskdef_ids = slits.maskdef_id
 
         # TODO: The above should set everything (self shouldn't be used
         # below). We need a single show method that both EdgeTraceSet
@@ -1369,6 +1391,7 @@ class EdgeTraceSet(DataContainer):
             id_kwargs = {'slit_ids': slit_ids} if synced \
                             else {'left_ids': traceid[gpm & is_left],
                                   'right_ids': traceid[gpm & is_right]}
+            id_kwargs['maskdef_ids'] = maskdef_ids
             _trc = cen if fit is None else fit
 
             # Connect to or instantiate ginga window
@@ -4021,6 +4044,7 @@ class EdgeTraceSet(DataContainer):
     
     def maskdesign_matching(self, debug=False):
         """
+        # TODO - break up this method (too long)
         Match slit info from the mask design data to the traced slits.
 
         Use of this method requires:
@@ -4043,6 +4067,10 @@ class EdgeTraceSet(DataContainer):
                 Run in debug mode.
         """
 
+        # Remove any fully masked traces. Keeps any inserted or
+        # box-slit traces.
+        self.clean_traces(rebuild_pca=True)
+
         # Check that there are traces to match!
         if self.is_empty:
             msgs.error('No traces to match.')
@@ -4052,11 +4080,12 @@ class EdgeTraceSet(DataContainer):
             msgs.error('Must first run the PCA analysis for the traces; run build_pca.')
 
         # `traceimg` must have knowledge of the flat frame that built it
-        if self.spectrograph.get_slitmask(self.traceimg.files[0]) is None:
+        self.maskfile = self.traceimg.files[0]
+        if self.spectrograph.get_slitmask(self.maskfile) is None:
             msgs.error('Unable to read slitmask design info')
-        if self.spectrograph.get_grating(self.traceimg.files[0]) is None:
+        if self.spectrograph.get_grating(self.maskfile) is None:
             msgs.error('Unable to read grating info')
-        if self.spectrograph.get_amapbmap(self.traceimg.files[0]) is None:
+        if self.spectrograph.get_amapbmap(self.maskfile) is None:
             msgs.error('Unable to read amap and bmap')
 
         # Match left and right edges separately
@@ -4230,7 +4259,7 @@ class EdgeTraceSet(DataContainer):
 
         # Find if there are missing traces.
         # Need exactly one occurrence of each index in "need"
-        buffer = 20.
+        buffer = self.par['det_buffer']+1
         need = ((top_edge_pred > buffer) & (bot_edge_pred < (self.traceimg.shape[1] - 1 - buffer))) & \
                ((omodel_bspat != -1) | (omodel_tspat != -1))
 
@@ -4269,6 +4298,17 @@ class EdgeTraceSet(DataContainer):
             ind_b = ind_b[np.logical_not(dupl_b)]
             ind_t = ind_t[np.logical_not(dupl_t)]
 
+        # RE-CHECK for missing traces after removing the duplicates
+        # bottom edges
+        needadd_b = need.copy()
+        needadd_b[ind_b] = False
+        needind_b = np.where(needadd_b)[0]  # edges we are missing
+
+        # top edges
+        needadd_t = need.copy()
+        needadd_t[ind_t] = False
+        needind_t = np.where(needadd_t)[0]  # edges we are missing
+
         #  The code below is to add traces that are predicted but not found.
         if needind_b.size > 0:
             msgs.info('Adding {} left missing edge(s)'.format(needind_b.size))
@@ -4277,29 +4317,64 @@ class EdgeTraceSet(DataContainer):
             sortind_b = np.argsort(utils.index_of_x_eq_y(self.spectrograph.slitmask.slitid[sortindx],
                                                          self.spectrograph.slitmask.slitid[ind_b], strict=True))
             ind_b = ind_b[sortind_b]
+            # check if any of the traces that will be added is off detector
+            loffdet = bot_edge_pred[needind_b] < self.par['det_buffer']
+            if np.any(loffdet):
+                # IF any, add a trace parallel to the detector edge (i.e., no pca)
+                self.insert_traces(-1, np.full(self.nspec, self.par['det_buffer']), mode='mask', nudge=False)
+                needind_b = needind_b[np.logical_not(loffdet)]
+            # We try to ensure that the left edge is inserted after the right edge of previous slit
+            for i in range(bot_edge_pred[needind_b].size):
+                # Find index of smaller closest trace
+                indx_smaller = np.where(self.edge_fit[self.pca.reference_row, :] < bot_edge_pred[needind_b][i])[0]
+                if (indx_smaller.size > 0) & (indx_smaller.size < self.traceid.size):
+                    indx = indx_smaller[-1]
+                    # If close trace is "left" and next one is "right", add the new trace after "right" one
+                    if (self.traceid[indx] < 0) & (self.traceid[indx+1] > 0) & \
+                            (self.edge_fit[self.pca.reference_row, indx + 1] - bot_edge_pred[needind_b][i] < 5):
+                        bot_edge_pred[needind_b[i]] = self.edge_fit[self.pca.reference_row, indx + 1] + 1
             # define which side to add the trace and insert it
             lside = -np.ones(bot_edge_pred[needind_b].shape[0], dtype=int)
+            # Predict trace
             missing_left_traces = self.predict_traces(bot_edge_pred[needind_b], side=lside)
-            self.insert_traces(lside, missing_left_traces, mode='mask')
+            # Add
+            self.insert_traces(lside, missing_left_traces, mode='mask', nudge=False)
 
         if needind_t.size > 0:
             msgs.info('Adding {} right missing edge(s)'.format(needind_t.size))
             # Append the missing indices and re-sort all
             ind_t = np.append(ind_t, needind_t)
             sortind_t = np.argsort(utils.index_of_x_eq_y(self.spectrograph.slitmask.slitid[sortindx],
-                                                         self.spectrograph.slitmask.slitid[ind_t],strict=True))
+                                                         self.spectrograph.slitmask.slitid[ind_t], strict=True))
             ind_t = ind_t[sortind_t]
+            # check if any of the traces that will be added is off detector
+            roffdet = top_edge_pred[needind_t] > self.nspat - self.par['det_buffer']
+            if np.any(roffdet):
+                # IF any, add a trace parallel to the detector edge (i.e., no pca)
+                self.insert_traces(1, np.full(self.nspec, self.nspat - (self.par['det_buffer'])),
+                                   mode='mask', nudge=False)
+                needind_t = needind_t[np.logical_not(roffdet)]
+            # We try to ensure that the left edge is inserted after the right edge of previous slit
+            for i in range(top_edge_pred[needind_t].size):
+                # Find index of larger closest trace
+                indx_larger = np.where(self.edge_fit[self.pca.reference_row, :] < top_edge_pred[needind_t][i])[0]
+                if (indx_larger.size > 0) & (indx_larger.size < self.traceid.size):
+                    indx = indx_larger[0]
+                    # If close trace is "right" and previous one is "left", add new trace before "left" one
+                    if (self.traceid[indx] > 0) & (self.traceid[indx-1] < 0) & \
+                            (top_edge_pred[needind_t][i] - self.edge_fit[self.pca.reference_row, indx - 1] < 5):
+                        top_edge_pred[needind_t[i]] = self.edge_fit[self.pca.reference_row, indx - 1] - 1
             # define which side to add the trace and insert it
             rside = np.ones(top_edge_pred[needind_t].shape[0], dtype=int)
+            # Predict trace
             missing_right_traces = self.predict_traces(top_edge_pred[needind_t], side=rside)
-            self.insert_traces(rside, missing_right_traces, mode='mask')
+            # Add
+            self.insert_traces(rside, missing_right_traces, mode='mask', nudge=False)
 
         if debug:
-            if (needind_b.size > 0) | np.any(dupl_b):
-                slitdesign_matching.plot_matches(self.edge_fit[:, self.is_left], ind_b, bot_edge_pred, reference_row,
+            slitdesign_matching.plot_matches(self.edge_fit[:, self.is_left], ind_b, bot_edge_pred, reference_row,
                                                  self.spectrograph.slitmask.slitindx, nspat=self.nspat, edge='left')
-            if (needind_t.size > 0) | np.any(dupl_t):
-                slitdesign_matching.plot_matches(self.edge_fit[:, self.is_right], ind_t, top_edge_pred, reference_row,
+            slitdesign_matching.plot_matches(self.edge_fit[:, self.is_right], ind_t, top_edge_pred, reference_row,
                                                  self.spectrograph.slitmask.slitindx, nspat=self.nspat, edge='right')
 
         self.maskdef_id = np.zeros(self.ntrace, dtype=int)
@@ -4312,7 +4387,6 @@ class EdgeTraceSet(DataContainer):
         align_slit[self.is_right] = self.spectrograph.slitmask.alignment_slit[ind_t]
         self.edge_msk[:, align_slit] = self.bitmask.turn_on(self.edge_msk[:, align_slit], 'BOXSLIT')
 
-
         # Propagate the coefficients, `coeff_b` and `coeff_t`, of the x-correlation and the
         # left and right spatial position of the slit edges from optical model (before x-correlation)
         # with the purpose to fill a table with the information on slitmask design matching. The table will
@@ -4322,7 +4396,19 @@ class EdgeTraceSet(DataContainer):
         self.omodel_bspat = omodel_bspat
         self.omodel_tspat = omodel_tspat
 
-    def _fill_design_table(self, maskdef_id, coeff_b, coeff_t, omodel_bspat, omodel_tspat):
+        # Make sure the traces are synchronized
+        if self.traceid[-1] < 0:
+            self.edge_msk[:, -1] = self.bitmask.turn_on(self.edge_msk[:, -1], 'OFFDETECTOR')
+        if self.traceid[0] > 0:
+            self.edge_msk[:, 0] = self.bitmask.turn_on(self.edge_msk[:, 0], 'OFFDETECTOR')
+        self.clean_traces(rebuild_pca=True)
+
+        if self.is_synced:
+            msgs.info('LEFT AND RIGHT EDGES SYNCHRONIZED AFTER MASK DESIGN MATCHING')
+        else:
+            msgs.warn('LEFT AND RIGHT EDGES *NOT* SYNCHRONIZED AFTER MASK DESIGN MATCHING')
+
+    def _fill_design_table(self, maskdef_id, coeff_b, coeff_t, omodel_bspat, omodel_tspat, spat_id):
         """
         Fill :attr:`design` based on the results of the design
         registration.
@@ -4356,6 +4442,8 @@ class EdgeTraceSet(DataContainer):
                 and traced edges for the left and right edges.
             omodel_bspat, omodel_tspat (:obj:`numpy.array`):
                 Left and right spatial position of the slit edges from optical model
+            spat_id (:obj:`numpy.array`):
+                ID assigned by PypeIt to each slit. same as in `SlitTraceSet`.
 
 
         """
@@ -4374,7 +4462,7 @@ class EdgeTraceSet(DataContainer):
         # Instantiate as an empty table
         self.design = EdgeTraceSet.empty_design_table(rows=nslits)
         # Save the fit parameters and the source file as table metadata
-        self.design.meta['MASKFILE'] = None
+        self.design.meta['MASKFILE'] = self.maskfile
         self.design.meta['MASKOFF'] = coeff_b[0], coeff_t[0]
         self.design.meta['MASKSCL'] = coeff_b[1], coeff_t[1]
         # Fill the columns
@@ -4385,6 +4473,7 @@ class EdgeTraceSet(DataContainer):
                                         dtype=self.design['TRACELPIX'].dtype)
         self.design['TRACERPIX'] = self.edge_fit[reference_row,self.traceid>0][matched].astype(
                                         dtype=self.design['TRACERPIX'].dtype)
+        self.design['SPAT_ID'] = spat_id[matched].astype(dtype=self.design['SPAT_ID'].dtype)
         self.design['SLITID'] = self.spectrograph.slitmask.slitid[ind].astype(
                                         dtype=self.design['SLITID'].dtype)
         self.design['SLITLOPT'] = omodel_bspat[ind].astype(dtype=self.design['SLITLOPT'].dtype)
@@ -4405,8 +4494,11 @@ class EdgeTraceSet(DataContainer):
         - 'OBJID': Object ID Number
         - 'OBJRA': Right ascension of the object (deg)
         - 'OBJDEC': Declination of the object (deg)
+        - 'OBJNAME': Object name assigned by the observer
         - 'SLITID': Slit ID Number (`maskdef_id`)
-        - 'SLITINDX': Row index of relevant slit in the design table
+        - 'OBJ_TOPDIST': Projected position of the object w.r.t. the top of the slit (arcsec)
+        - 'OBJ_BOTDIST': Projected position of the object w.r.t. the bottom of the slit (arcsec)
+        - 'TRACEID': Row index that matches 'TRACEID' in the design table
 
         Args:
             maskdef_id (:obj:`numpy.array`):
@@ -4423,7 +4515,7 @@ class EdgeTraceSet(DataContainer):
         obj_index = utils.index_of_x_eq_y(self.spectrograph.slitmask.objects[:,0], maskdef_id, strict=False)
         # if not all the element of self.spectrograph.slitmask.objects[obj_index,0] are equal to maskdef_id,
         # keep only the elements that are equal (matched)
-        matched = np.where(self.spectrograph.slitmask.objects[obj_index,0] == maskdef_id)[0]
+        matched = np.where(self.spectrograph.slitmask.objects[obj_index,0].astype(int) == maskdef_id)[0]
         obj_index = obj_index[matched]
 
         # Number of objects
@@ -4431,17 +4523,13 @@ class EdgeTraceSet(DataContainer):
         # Instantiate an empty table
         self.objects = EdgeTraceSet.empty_objects_table(rows=nobj)
         # Fill the columns
-        for i,key in enumerate(['SLITID', 'OBJID', 'OBJRA', 'OBJDEC']):
-                self.objects[key] = self.spectrograph.slitmask.objects[obj_index,i].astype(
-                                        dtype=self.objects[key].dtype)
-        #TODO it would be good to add also the 'OBJECT' keyword from the slit-mask design.
-        # 'OBJECT' is the name that the observer gives to each target and therefore more easily recognizable
+        for i,key in enumerate(['SLITID', 'OBJID', 'OBJRA', 'OBJDEC', 'OBJNAME', 'OBJ_TOPDIST', 'OBJ_BOTDIST']):
+            self.objects[key] = self.spectrograph.slitmask.objects[obj_index,i].astype(dtype=self.objects[key].dtype)
 
         # SLITINDX is the index of the slit in the `design` table, not
         # in the original slit-mask design data
-        self.objects['SLITINDX'] = utils.index_of_x_eq_y(self.objects['SLITID'],
+        self.objects['TRACEID'] = utils.index_of_x_eq_y(self.objects['SLITID'],
                                                          self.design['SLITID'], strict=True)
-
 
 # NOTE: I'd like us to keep this commented mask_refine function around
 # for the time being.
@@ -4978,6 +5066,11 @@ class EdgeTraceSet(DataContainer):
             specmin = specmin[indx]/binspec
             specmax = specmax[indx]/binspec
 
+        # Define spat_id (in the same way is done in SlitTraceSet) to add it in the astropy table. It
+        # is useful to have spat_id in that table.
+        center = (left + right) / 2
+        spat_id = np.round(center[int(np.round(0.5 * self.nspec)), :]).astype(int)
+
         if self.maskdef_id is not None:
             # Check for mismatched `maskdef_id` in the left and right edges
             mkd_id_mismatch = self.maskdef_id[self.is_left] != self.maskdef_id[self.is_right]
@@ -4995,21 +5088,26 @@ class EdgeTraceSet(DataContainer):
                           "They will not be included in the design table")
 
             # Store the matched slit-design and object information in a table.
-            self._fill_design_table(_maskdef_id, self.coeff_b, self.coeff_t, self.omodel_bspat, self.omodel_tspat)
+            self._fill_design_table(_maskdef_id, self.coeff_b, self.coeff_t, self.omodel_bspat,
+                                    self.omodel_tspat, spat_id)
             self._fill_objects_table(_maskdef_id)
-            _merged_designtab = table.join(self.design, self.objects, keys=['SLITID'])
-            _merged_designtab.sort('TRACEID')
-            _merged_designtab.remove_column('SLITINDX')
+            # TODO - instead of merging the two tables, just create a single one
+            _merged_designtab = table.join(self.design, self.objects, keys=['TRACEID'])
+            _merged_designtab.remove_column('SLITID_2')
+            _merged_designtab.rename_column('SLITID_1', 'SLITID')
+            # One more item
+            _posx_pa = float(self.spectrograph.slitmask.posx_pa)
         else:
             _maskdef_id = None
             _merged_designtab = None
+            _posx_pa = None
 
         # Instantiate and return
-        return slittrace.SlitTraceSet(left, right, self.spectrograph.pypeline, nspat=self.nspat,
+        return slittrace.SlitTraceSet(left, right, self.spectrograph.pypeline,
+                                      det=self.traceimg.detector.det, nspat=self.nspat,
                                       PYP_SPEC=self.spectrograph.name, specmin=specmin,
                                       specmax=specmax, binspec=binspec, binspat=binspat,
                                       pad=self.par['pad'], mask_init=slit_msk,
                                       maskdef_id=_maskdef_id, maskdef_designtab=_merged_designtab,
+                                      maskdef_posx_pa=_posx_pa, maskfile=self.maskfile,
                                       ech_order=ech_order)
-
-