+++ CHANGES.rst
@@ -69,6 +69,12 @@
       problem.
 - Kastr 300 grating solutions
 - Hotfix to include the solutions!
+- Improved DEIMOS slitmask design matching
+- Assign RA/DEC to DEIMOS extractions
+- DEIMOS object RA, Dec, and name returned when running `pypeit_show_1d --list` and saved in
+  the .txt file with the list of 1d spectra.
+- DEIMOS object name and `maskdef_id` visible in ginga when running `pypeit_show_2d`
+- Fix sigma clipping bug!
 
 1.2.0 (15 Oct 2020)
 -------------------